requirements.txt
geoip2 >= 4.4.0
PyYAML >= 6.0
requests >= 2.28.1
psutil >= 5.9.4